Add files via upload
This radio is utterly an educational attempt on my part to build a web radio and learn something new. I am not a professional programmer so please be patient and bear with my coding "style". If you are looking for some really good piece of project, please refer to somewhere else. 
Drawings have been inspired to the super project by https://github.com/Edzelf. At the time of writing I am stuck with BBC channels  not working well since when in past February they changed their streams. There is always a nasty glitch which I am not able to get rid of. 
I have bee trying with adopting a ring buffer and/or relying  on rtos queues and semaphores with no success. I have been stripping and studying Edzelf code as well but where his code works, mine fails miserably. This is why in this version I decided to remain cling to the simplest "read stream and feed it to the mp3 chip"
Characteristics:
1) Works with low bit rate stations;
2) Supports IR remote controls: Apple, Elegoo and my home's Samsung tv;
3) OLED 1306 (128x64);
4) Mp3 through a VS1053.

To do:
1) Fix the BBC radios hassle;
2) As 1) but with Otto Baroque Radio (less nasty then n.1)